requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pyYaml
python-box==6.0.2
tqdm
ensure==1.0.2
joblib
dataclasses
requests
imblearn
jupyterlab

-e .
